Move SAM2 detector into separate file to avoid dependency conflicts
